some refactoring and WIP subs
