+++ src/types/volleyballTool.New.Types.ts
@@ -515,3 +515,69 @@ export const RotationPositions: RotationPositionType = {
   },
 };
 //#endregion
+
+//#region Positions
+type PositionType = {
+  id: string;
+  name: string;
+  shortName?: string;
+};
+
+//TODO: move to a new folder called static data (along with enums)
+export const OutsideHitter: PositionType = {
+  id: "1",
+  name: "Outside Hitter",
+  shortName: "OH",
+};
+
+export const OppositeHitter: PositionType = {
+  id: "2",
+  name: "Opposite Hitter",
+  shortName: "Oppo",
+};
+
+export const MiddleBlocker: PositionType = {
+  id: "3",
+  name: "Middle Blocker",
+  shortName: "MB",
+};
+
+export const Setter: PositionType = {
+  id: "4",
+  name: "Setter",
+  shortName: "Set",
+};
+
+export const Libero: PositionType = {
+  id: "5",
+  name: "Libero",
+  shortName: "Lib",
+};
+
+export const None: PositionType = {
+  id: "6",
+  name: "None",
+  shortName: "None",
+};
+
+export const PositionsById = {
+  [OutsideHitter.id]: OutsideHitter,
+  [MiddleBlocker.id]: MiddleBlocker,
+  [Setter.id]: Setter,
+  [OppositeHitter.id]: OppositeHitter,
+  [Libero.id]: Libero,
+  [None.id]: None,
+};
+
+/**
+ * Position type (major positions only)
+ */
+export type Position =
+  | typeof OutsideHitter
+  | typeof OppositeHitter
+  | typeof MiddleBlocker
+  | typeof Setter
+  | typeof Libero
+  | typeof None;
+
+//#endregion
